requirements.txt
absl-py==0.12.0
beautifulsoup4==4.9.3
colorlog==4.8.0
Cython==0.29.24
descartes==1.1.0
einops==0.3.0
future==0.18.2
GPUtil==1.4.0
h5py==3.1.0
hydra-colorlog==1.0.0
hydra-core==1.0.6
imageio==2.9.0
matplotlib
numpy
oiio @ https://github.com/fredrikaverpil/oiio-python/releases/download/2.0.5%2B20190203/oiio-2.0.5-cp36-none-linux_x86_64.whl
omegaconf==2.0.6
open3d==0.9.0.0
opencv-python==4.5.1.48
openmesh==1.1.6
opt-einsum==3.3.0
plyfile==0.7.3
projectq==0.3.6
protobuf==3.15.6
pyquaternion==0.9.9
pytransform3d==1.7.1
pyvista==0.31.0
PyYAML==5.4.1
sklearn==0.0
tqdm==4.32.1
transforms3d==0.3.1
trimesh==3.9.10
tvb-gdist==2.0.2
wandb==0.10.23
yacs==0.1.8
contextlib2
lmdb
gdown
easydict
dataclasses
pytest
pybind11
requests
scikit-image
scikit-learn
Pillow
six
pyrender
